pATCH

+++ com.luckia.biller.deploy/src/main/java/com/luckia/biller/deploy/Patch20140522B.java
@@ -22,8 +22,8 @@ import com.luckia.biller.core.services.bills.LiquidationProcessor;
 public class Patch20140522B {
 
 	private static final Logger LOG = LoggerFactory.getLogger(Main.class);
-	private static final boolean PROCESS_STORES = false;
-	private static final boolean PROCESS_COMPANIES = true;
+	private static final boolean DELETE = false;
+	private static final boolean CREATE = true;
 
 	public static void main(String[] args) {
 		LOG.info("Ejecutando patch");
@@ -35,24 +35,28 @@ public class Patch20140522B {
 		BillProcessor billProcessor = injector.getInstance(BillProcessor.class);
 		LiquidationProcessor liquidationProcessor = injector.getInstance(LiquidationProcessor.class);
 		EntityManager entityManager = entityManagerProvider.get();
-		String[] ids = { "17cc6828-c27a-4850-9734-8302701f4715", "5533edd6-91c4-4f37-9972-6e8ae2932c95" };
-		for (String liquidationId : ids) {
-			Liquidation liquidation = entityManagerProvider.get().find(Liquidation.class, liquidationId);
-			entityManager.getTransaction().begin();
-			LOG.info("Eliminando relacion de facturas con la liquidacion");
-			for (Bill bill : liquidation.getBills()) {
-				bill.setLiquidation(null);
-				entityManager.merge(bill);
-			}
-			entityManager.getTransaction().commit();
+		if (DELETE) {
+			String[] ids = { "17cc6828-c27a-4850-9734-8302701f4715", "5533edd6-91c4-4f37-9972-6e8ae2932c95" };
+			for (String liquidationId : ids) {
+				Liquidation liquidation = entityManagerProvider.get().find(Liquidation.class, liquidationId);
+				entityManager.getTransaction().begin();
+				LOG.info("Eliminando relacion de facturas con la liquidacion");
+				for (Bill bill : liquidation.getBills()) {
+					bill.setLiquidation(null);
+					entityManager.merge(bill);
+				}
+				entityManager.getTransaction().commit();
 
-			LOG.info("Eliminando liquidacion");
-			entityManager.getTransaction().begin();
-			entityManager.remove(liquidation);
-			entityManager.getTransaction().commit();
+				LOG.info("Eliminando liquidacion");
+				entityManager.getTransaction().begin();
+				entityManager.remove(liquidation);
+				entityManager.getTransaction().commit();
+			}
 		}
-		// long companyId = 168;
-		// LiquidationTask task = new LiquidationTask(companyId, range, entityManagerProvider, liquidationProcessor);
-		// task.run();
+		if (CREATE) {
+			long companyId = 168;
+			LiquidationTask task = new LiquidationTask(companyId, range, entityManagerProvider, liquidationProcessor);
+			task.run();
+		}
 	}
 }
